Server shares upload limit with client
When uploading a file, the client sees the upload limit for the server

+++ units/units.go
@@ -1,10 +1,10 @@
 package units
 
 //1 kilobyte (1024 bytes)
-var KB int16 = 1024
+var KB uint16 = 1024
 
 //1 megabyte (1048576 bytes)
-var MB int32 = 1048576
+var MB uint32 = 1048576
 
 //1 gigabyte (1073741824 bytes)
-var GB int32 = 1073741824
+var GB int64 = 1073741824
